Some pretty major changes
Added a new class information file to help keep track of the classes methods and variables. Also made a much simpler way to track flags bombs counted tiles and empty tiles eliminating the need for the Block class. Also created a method to clear a square around a selected tile showing the ones around it on the cover board. Otherwise known as the user board or the board that the user will see. The GUI